fix: change server api base and cover with tests
