+++ src/controllers/settings.controller.ts
@@ -3,7 +3,7 @@ import { authenticate, authorizeRoles } from "@/middleware/authenticate";
 import { AppConstants } from "@/server.constants";
 import { ROLES } from "@/constants/authorization.constants";
 import { validateInput } from "@/handlers/validators";
-import { whitelistSettingRules, sentryDiagnosticsEnabledRules } from "./validation/setting.validation";
+import { whitelistSettingRules, sentryDiagnosticsEnabledRules } from "@/services/validators/settings-service.validation";
 import { SettingsStore } from "@/state/settings.store";
 import { Request, Response } from "express";
 import { IpWhitelistSettingsDto } from "@/services/interfaces/settings.dto";
@@ -53,8 +53,8 @@ export class SettingsController {
 export default createController(SettingsController)
   .prefix(AppConstants.apiRoute + "/settings")
   .before([authenticate(), authorizeRoles([ROLES.ADMIN])])
-  .get("/server", "getSettings")
-  .put("/server/server", "updateServerSettings")
-  .patch("/server/sentry-diagnostics", "updateSentryDiagnosticsEnabled")
-  .put("/server/whitelist", "updateWhitelistSettings")
-  .put("/server/frontend", "updateFrontendSettings");
+  .get("/", "getSettings")
+  .put("/server", "updateServerSettings")
+  .patch("/sentry-diagnostics", "updateSentryDiagnosticsEnabled")
+  .put("/whitelist", "updateWhitelistSettings")
+  .put("/frontend", "updateFrontendSettings");
